feat(back): improve both the date validations in the date validation folder as well as every function from the controllers folder to get a better performance

+++ back/src/api/controllers/attendee.js
@@ -18,13 +18,13 @@ const getAttendeeById = async (req, res, next) => {
         const { id } = req.params;
         const attendee = await Attendee.findById(id).populate({ path: 'username', select: 'username' }).populate({ path: 'email', select: 'email' }).populate({ path: 'attended_events', select: 'title' });
         if (!attendee) {
-            const error = new Error("the attendee couldn't be found");
+            const error = new Error(`the attendee couldn't be found`);
             error.statusCode = 404;
             return next(error);
         }
         return res.status(200).json(attendee);
     } catch (err) {
-        const error = new Error("an error occurred displaying the attendee's data");
+        const error = new Error(`an error occurred displaying the attendee's data`);
         error.statusCode = 500;
         next(error);
     }
@@ -32,11 +32,18 @@ const getAttendeeById = async (req, res, next) => {
 
 const postAttendanceToAnEvent = async (req, res, next) => {
     try {
-        const { id } = req.params;
+        const userId = req.params.id;
+        const eventId = req.params.eventId;
 
-        if (req.user.organized_events.some(event => event._id.toString() === id)) {
-            const error = new Error("the user can't confirm attendance because of being the event host");
+        if (req.user.organized_events.some(event => event._id.toString() === eventId)) {
+            const error = new Error(`the user can't confirm attendance because of being the event host`);
             error.statusCode = 400;
+            return next(error);
+        }
+
+        if (req.user.role === 'user' && userId !== req.user._id.toString()) {
+            const error = new Error(`It's not allowed to post attendance at another user's event.`);
+            error.statusCode = 403;
             return next(error);
         }
 
@@ -55,45 +62,52 @@ const postAttendanceToAnEvent = async (req, res, next) => {
         const savedNewAttendee = await newAttendee.save();
 
         await Attendee.findByIdAndUpdate(savedNewAttendee._id, {
-            $push: { attended_events: { _id: id } }
+            $push: { attended_events: { _id: eventId } }
         });
 
-        await Event.findByIdAndUpdate(id, {
-            $push: { attendees: { _id: req.user._id } }
+        await Event.findByIdAndUpdate(eventId, {
+            $push: { attendees: { _id: userId } }
         });
 
-        await User.findByIdAndUpdate(req.user._id, {
-            $push: { attended_events: { _id: id } }
+        await User.findByIdAndUpdate(userId, {
+            $push: { attended_events: { _id: eventId } }
         });
 
         return res.status(201).json(savedNewAttendee);
     } catch (err) {
         const error = new Error('an error occurred confirming attendance to the event');
         error.statusCode = 500;
-        next(error);
+        return next(error);
     }
 };
 
 const deleteAttendanceToAnEvent = async (req, res, next) => {
     try {
-        const { id } = req.params;
+        const userId = req.params.id;
+        const eventId = req.params.eventId;
 
         const attendee = await Attendee.findOne({ username: req.user.username });
 
         if (!attendee) {
-            const error = new Error("the attendee couldn't be found");
+            const error = new Error(`the attendee couldn't be found`);
             error.statusCode = 404;
             return next(error);
         }
 
-        await Event.findByIdAndUpdate(id, { $pull: { attendees: req.user._id } });
+        if (req.user.role === 'user' && userId !== req.user._id.toString()) {
+            const error = new Error(`it's not allowed to delete another user's event attendance`);
+            error.statusCode = 403;
+            return next(error);
+        }
 
-        await User.findByIdAndUpdate(req.user._id, { $pull: { attended_events: id } });
+        await Event.findByIdAndUpdate(eventId, { $pull: { attendees: userId } });
+
+        await User.findByIdAndUpdate(userId, { $pull: { attended_events: eventId } });
 
         const deletedAttendee = await Attendee.findByIdAndDelete(attendee._id);
         return res.status(200).json(deletedAttendee);
     } catch (err) {
-        const error = new Error("an error occurred deleting the attendee's data");
+        const error = new Error(`an error occurred deleting the attendee's data`);
         error.statusCode = 500;
         next(error);
     }
